Moved OrgUser methods to separate file
